// src/models/ImageDetailsDetailed.ts
// Generated by https://quicktype.io

export interface ImageDetailsDetailed {
    Make:                      string;
    Model:                     string; //
    LensModel:                 string; //
    Orientation:               string;
    XResolution:               number;
    YResolution:               number;
    ResolutionUnit:            string;
    Software:                  string;
    YCbCrPositioning:          number;
    ExposureTime:              number; //
    FNumber:                   number; //
    ExposureProgram:           string;
    ISO:                       number; //
    SensitivityType:           number;
    StandardOutputSensitivity: number;
    ExifVersion:               string;
    ShutterSpeedValue:         number;
    ApertureValue:             number;
    ExposureCompensation:      number;
    MeteringMode:              string;
    Flash:                     string;
    FocalLength:               number; //
    FlashpixVersion:           string;
    ColorSpace:                number;
    ExifImageWidth:            number;
    ExifImageHeight:           number;
    SensingMethod:             string;
    FileSource:                string;
    SceneType:                 string;
    CustomRendered:            string;
    ExposureMode:              string;
    WhiteBalance:              string;
    FocalLengthIn35mmFormat:   number; //
    SceneCaptureType:          string;
    Contrast:                  string;
    Saturation:                string;
    Sharpness:                 string;
    SubjectDistanceRange:      string;
    GPSVersionID:              string;
    FocalPlaneXResolution:     number;
    FocalPlaneYResolution:     number;
    FocalPlaneResolutionUnit:  string;

    DateTimeOriginal: Date|string; //
    FileName: string; //
    FileType: string; //
    FileSize: number, //
    [key:string]: any
    
}